Fix crash in multithreaded mode
Application instantly crashes in new EPICS V3.15, that utilizes multithreaded
design. The reason for this is that devSysMon.c was designed for old single-thread
mode. It uses the same variables to store file descriptor and read buffer for
several independent groups of PVs. In multithreaded mode concurrent processing
of PVs leads to simultaneous use the mentioned variables, which causes crash.
To solve the problem the code was improved to fully separate variables for
independent groups.