Add disk and network I/O measurements
Since there can be many disks and interfaces on a single system,
one has to instantiate pattern mode for DB generation. Downside
is that one can't simply use FLNK field to chain PVs as before.
For the moment I decided to use PHAS to put processing before
read of values. Yet this method is called unreliable in the manual
I don't see difficulties at the moment.